+++ cmd/generate.go
@@ -10,8 +10,11 @@ import (
 )
 
 var (
-	mask      string
-	minLength int
+	mask              string
+	minLength         int
+	extractHcPotfiles string
+	permutationFile   string
+	maxLength         int
 )
 
 var generateCmd = &cobra.Command{
@@ -22,45 +25,67 @@ var generateCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		outputPath := args[0]
 
-		outputPath, err := utils.OverwriteFile(outputPath)
+		outputPath, err := utils.OverwriteFile(outputPath) // checks if output file is existing
 		if err != nil {
 			output.PrintError("errors", "error", err)
 			os.Exit(1)
 		}
-		stop := make(chan struct{})
+
+		stop := make(chan struct{}) // spinner animation
 		go output.Spinner("Generate File", stop)
 
-		// when generate is called
-		if minLength == 0 {
-			entities, bytes, err := generator.CalculateMaskStorage(mask)
+		if extractHcPotfiles != "" { // hashcat password extraction is chosen
+			err := generator.ExtractHashCatPotfile(extractHcPotfiles, outputPath)
 			if err != nil {
 				output.PrintError("errors", "error", err)
 				os.Exit(1)
 			}
-			output.PrintStatus("statusGenerator", "calculateWords", utils.FormatUint64WithCommas(entities))
-			output.PrintStatus("statusGenerator", "calculateSize", utils.HumanReadableBytes(bytes))
-			output.PrintStatus("statusGenerator", "buildingMaskWordlist", mask)
+		} else if mask != "" { // mask generation is chosen
+			// when generate is called
+			if minLength == 0 {
+				entities, bytes, err := generator.CalculateMaskStorage(mask)
+				if err != nil {
+					output.PrintError("errors", "error", err)
+					os.Exit(1)
+				}
+				output.PrintStatus("statusGenerator", "calculateWords", utils.FormatUint64WithCommas(entities))
+				output.PrintStatus("statusGenerator", "calculateSize", utils.HumanReadableBytes(bytes))
+				output.PrintStatus("statusGenerator", "buildingMaskWordlist", mask)
 
-			err = generator.ConcurrentGenerateMaskWordlist(mask, outputPath)
-			if err != nil {
-				output.PrintError("errors", "error", err)
-				os.Exit(1)
-			}
-		} else {
-			entities, bytes, err := generator.CalculateMaskStorage(mask)
-			if err != nil {
-				output.PrintError("errors", "error", err)
-				os.Exit(1)
-			}
-			output.PrintStatus("statusGenerator", "calculateWords", utils.FormatUint64WithCommas(entities))
-			output.PrintStatus("statusGenerator", "calculateSize", utils.HumanReadableBytes(bytes))
-			output.PrintStatus("statusGenerator", "buildingMaskWordlist", mask)
-
-			err = generator.ConcurrentGenerateMaskWordlist(mask, outputPath, minLength)
-			if err != nil {
-				output.PrintError("errors", "error", err)
-				os.Exit(1)
+				err = generator.ConcurrentGenerateMaskWordlist(mask, outputPath)
+				if err != nil {
+					output.PrintError("errors", "error", err)
+					os.Exit(1)
+				}
 			} else {
+				entities, bytes, err := generator.CalculateMaskStorage(mask)
+				if err != nil {
+					output.PrintError("errors", "error", err)
+					os.Exit(1)
+				}
+				output.PrintStatus("statusGenerator", "calculateWords", utils.FormatUint64WithCommas(entities))
+				output.PrintStatus("statusGenerator", "calculateSize", utils.HumanReadableBytes(bytes))
+				output.PrintStatus("statusGenerator", "buildingMaskWordlist", mask)
+
+				err = generator.ConcurrentGenerateMaskWordlist(mask, outputPath, minLength)
+				if err != nil {
+					output.PrintError("errors", "error", err)
+					os.Exit(1)
+				} else {
+					output.PrintError("errors", "error", err)
+					os.Exit(1)
+				}
+			}
+		} else if permutationFile != "" && maxLength != 0 { // permutation chosen
+			if maxLength <= 0 {
+				output.PrintError("errors", "wrongMaxLength", "")
+				os.Exit(1)
+			}
+			if minLength > maxLength && maxLength < 0 {
+				output.PrintError("errors", "wrongMinLength", "")
+			}
+			err := generator.ConcurrentPermutation(permutationFile, outputPath, minLength, maxLength, 0)
+			if err != nil {
 				output.PrintError("errors", "error", err)
 				os.Exit(1)
 			}
@@ -70,7 +95,7 @@ var generateCmd = &cobra.Command{
 			output.PrintError("errors", "error", err)
 			os.Exit(1)
 		}
-		close(stop)
+		close(stop) // stopping spinner animation
 		output.PrintSuccess("successGenerator", "wordlistCreated", oPath+utils.ListSuffix)
 	},
 }
@@ -79,11 +104,10 @@ func init() {
 	rootCmd.AddCommand(generateCmd)
 
 	generateCmd.Flags().StringVarP(&mask, "mask", "M", "", output.GenerateMessages["mask"])
-	err := generateCmd.MarkFlagRequired("mask")
-	if err != nil {
-		output.PrintError("errors", "error", err)
-		os.Exit(1)
-	}
-	generateCmd.Flags().IntVarP(&minLength, "min-length", "m", 0, output.GenerateMessages["minLength"])
+
+	generateCmd.Flags().IntVar(&minLength, "min", 0, output.GenerateMessages["minLength"])
+	generateCmd.Flags().StringVarP(&permutationFile, "permutation", "p", "", output.GenerateMessages["permutation"])
+	generateCmd.Flags().IntVarP(&maxLength, "max", "m", 0, output.GenerateMessages["maxLength"])
+	generateCmd.Flags().StringVar(&extractHcPotfiles, "extract-hc-potfile", "", output.GenerateEditHelpTexts["extract-hc-potfile"])
 
 }