+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Please Contact me if you find typos.
 The format is based on [Keep a Changelog](https://keepachangelog.com)       
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+## upcoming
+- added permutation wordlist generation
+- hashcat potfile extraction is now in generate instead of edit
+
 ## koyane-framework 0.7.0-beta
 - Complete program remastered in GO
 - All computationally intensive algorithms are parallelized.